fix(brokerage): moving ineligible error message text to modal brokerage components dir
