+++ packages/blockchain-wallet-v4-frontend/src/components/DataError/ErrorHandler.js
@@ -1,7 +1,7 @@
 import {
   BROKERAGE_INELIGIBLE,
   IneligibleErrorMessage
-} from 'components/Brokerage'
+} from '../../modals/Brokerage/components'
 import { Button, Link, Text, TextGroup } from 'blockchain-info-components'
 import { checkForVulnerableAddressError } from 'services/ErrorCheckService'
 import { FETCH_FEES_FAILURE } from 'blockchain-wallet-v4/src/redux/payment/model'
